requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alpha_vantage==3.0.0
altair==5.5.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
captcha==0.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.2.0
extra-streamlit-components==0.1.71
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
holidays==0.66
html5lib==1.1
idna==3.10
investpy==1.0.8
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
narwhals==1.26.0
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
streamlit==1.42.0
streamlit-authenticator==0.4.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
yarl==1.18.3
yfinance==0.2.54